[ARVIND] : scenario-> Added the program file of employee using java8 stream

+++ java8-programmming-practice/scenario-code-base/ScenarioBased/src/com/questions/employee/Employee.java
@@ -6,11 +6,10 @@ public class Employee {
 	private int age;
 	private String gender;
 	private String department;
-	private double yearOfJoining;
+	private int yearOfJoining;
 	private double salary;
 	
-	public Employee(int id, String name, int age, String gender, String department, double yearOfJoining,
-			double salary) {
+	public Employee(int id, String name, int age, String gender, String department, int yearOfJoining, double salary) {
 		this.id = id;
 		this.name = name;
 		this.age = age;
@@ -19,63 +18,49 @@ public class Employee {
 		this.yearOfJoining = yearOfJoining;
 		this.salary = salary;
 	}
-
+	
 	public int getId() {
 		return id;
 	}
-
 	public void setId(int id) {
 		this.id = id;
 	}
-
 	public String getName() {
 		return name;
 	}
-
 	public void setName(String name) {
 		this.name = name;
 	}
-
 	public int getAge() {
 		return age;
 	}
-
 	public void setAge(int age) {
 		this.age = age;
 	}
-
 	public String getGender() {
 		return gender;
 	}
-
 	public void setGender(String gender) {
 		this.gender = gender;
 	}
-
 	public String getDepartment() {
 		return department;
 	}
-
 	public void setDepartment(String department) {
 		this.department = department;
 	}
-
-	public double getYearOfJoining() {
+	public int getYearOfJoining() {
 		return yearOfJoining;
 	}
-
-	public void setYearOfJoining(double yearOfJoining) {
+	public void setYearOfJoining(int yearOfJoining) {
 		this.yearOfJoining = yearOfJoining;
 	}
-
 	public double getSalary() {
 		return salary;
 	}
-
 	public void setSalary(double salary) {
 		this.salary = salary;
 	}
-
 	@Override
 	public String toString() {
 		return "Employee [id=" + id + ", name=" + name + ", age=" + age + ", gender=" + gender + ", department="
